fix(fzf): homebrew prefix var & iterm2 keybind
in Iterm2, by default alt acts like mac os option, which is used to
create international characters. This interferes with any keybinds that
use alt, naturally, like alt+c, which is used by fzf to change
directory.

By binding the international character to that function in ZLE, it
works even if the terminal hasn't been set to use Esc+ as the input for alt.