fix(emacs): load org config after module
The org config wasn't loading in the right order, and was getting
overwritten by the doom org module.

This also changes the daily capture to target the journal heading, which
is now in the daily template, and adds the MARV todo state for when I've
moved a todo to marvin.